feat(form)!: replace SelectField with the declarative OptionsField API
BREAKING CHANGE: SelectField no longer takes items/getKey/getLabel/getValue.
Options are now written directly as <Option value={...}> children (what was
OptionsField). Combobox is unaffected — it keeps the items array API for
searchable/large/dynamic lists. Update the internal fields demo page to match.
