+++ src/lib/shared/components/form/fields/select-field-context.ts
@@ -1,0 +1,17 @@
+import type { Snippet } from 'svelte';
+
+/** Shared context key linking a declarative `<Option>` to its ancestor `<SelectField>`. */
+export const SELECT_FIELD_CONTEXT = Symbol('web-svelte:select-field');
+
+export type SelectFieldEntry = {
+  content: Snippet;
+  label: string;
+  /** The option's typed value — read back as `unknown`, cast to `V` by `<SelectField>` (the
+   *  only place that knows the concrete `V` for this instance). */
+  value: unknown;
+};
+
+export type SelectFieldContext = {
+  register: (key: string, entry: SelectFieldEntry) => void;
+  unregister: (key: string) => void;
+};
